Part-A task 2:modified files to allow only owners to edit/delete images

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -94,14 +94,25 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
+    //The method makes sure to allow only the user to edit the images and redirects accordingly
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
+        User user = (User) session.getAttribute("loggeduser");
         model.addAttribute("image", image);
+        String tags = convertTagsToString(image.getTags());
         model.addAttribute("tags", tags);
-        return "images/edit";
+
+        if(user.getId() == image.getUser().getId())
+            return "images/edit";
+        else
+        {
+            String error = "Only the owner of the image can edit the image";
+            List<Tag> tagList = findOrCreateTags(tags);
+            model.addAttribute("tags",tagList);
+            model.addAttribute("editError",error);
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -135,17 +146,33 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" +updatedImage.getId()+"/"+ updatedImage.getTitle();
     }
 
 
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
-    //Looks for a controller method with request mapping of type '/images'
+    //The method makes sure to allow only the user to delete the images  and redirects accordingly
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        User user = (User) session.getAttribute("loggeduser");
+        String tags = convertTagsToString(image.getTags());
+
+        if(user.getId() == image.getUser().getId())
+        {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else
+        {
+            String error = "Only the owner of the image can delete the image";
+            List<Tag> tagList = findOrCreateTags(tags);
+            model.addAttribute("image",image);
+            model.addAttribute("tags",tagList);
+            model.addAttribute("deleteError",error);
+            return "images/image";
+        }
     }
